change position file
