+++ src/pages/common/IndexMailSend.js
@@ -1,7 +1,7 @@
 import { Component } from "react";
-import MenuMail from "../../components/employer/mail/MenuMail";
-import GetMail from "../../components/employer/mail/GetMail";
-class DetaiMail extends Component {
+import MenuMail from "../../components/common/mail/MenuMail";
+import ListMailSend from "../../components/common/mail/ListMailSend";
+class IndexMailSend extends Component {
   render() {
     document.body.style.backgroundColor = "#eceff1 ";
     return (
@@ -9,10 +9,10 @@ class DetaiMail extends Component {
         <h1 className="title-email">Hộp thư</h1>
         <div className="row">
           <MenuMail />
-          <GetMail id={this.props.id} />
+          <ListMailSend />
         </div>
       </div>
     );
   }
 }
-export default DetaiMail;
+export default IndexMailSend;
